requirements.txt
# Core Backend
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.9.0
python-dotenv>=1.0.1

# AI & Orchestration
langchain>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
openai>=1.50.0
qdrant-client>=1.12.0
# voyageai>=0.3.0  <-- Commented out for local dev
ollama>=0.4.0
fastembed>=0.4.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Ingestion & Scraping
httpx>=0.27.0
beautifulsoup4>=4.12.0
feedparser>=6.0.11
firecrawl-py>=0.0.20
playwright>=1.40.0
playwright-stealth>=2.0.1
html2text>=2024.2.26

# UI
streamlit>=1.40.0

# Utils
loguru>=0.7.0